requirements.txt
astroid==2.15.6
beautifulsoup4==4.12.2
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
dill==0.3.6
distlib==0.3.6
filelock==3.12.2
greenlet==2.0.2
identify==2.5.24
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
nodeenv==1.8.0
numpy==1.25.1
pandas==2.0.3
platformdirs==3.8.1
pre-commit==3.3.3
pylint==2.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
wrapt==1.15.0
